third_party/WebKit/LayoutTests/svg/animations/svgnumber-animation-3-expected.txt

CONSOLE WARNING: SVG's SMIL animations (<animate>, <set>, etc.) are deprecated and will be removed. Please use CSS animations or Web animations instead.
SVG 1.1 dynamic animation tests

Test for SVGNumber animation with invalid units.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getComputedStyle(rect).opacity is '0'
PASS getComputedStyle(rect).opacity is '0'
PASS getComputedStyle(rect).opacity is '0'
PASS getComputedStyle(rect).opacity is '0'
PASS successfullyParsed is true

TEST COMPLETE
